Improve of performance and code

+++ epochs/hippieeg/code/epochs.py
@@ -3,37 +3,20 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import traceback
-from file_manager import create_EDF, find
+from file_manager import create_EDF, create_bipolars, create_bipolar_tsv
 import argparse
 import re
 import shutil
 import os
 
-parser = argparse.ArgumentParser(description='Script to extract epochs\
-                                             from a *.edf file.')
-parser.add_argument("-p", "--processes" ,metavar='PROCESESS',
-                    nargs = 1,
-                    help='Number of processes per patient.')
-parser.add_argument("-i", "--input", metavar='INPUT',
-                    nargs = '*',
-                    help='Path to subject directory')
-parser.add_argument("-a", "--annot", metavar='ANNOTATION',
-                    nargs = '*',
-                    help='Path to annotation file')
-parser.add_argument("-t", "--tsv", metavar='INPUT_TSV',
-                    nargs = '*',
-                    help='Path to TSV file')
-parser.add_argument("-o", "--output" ,metavar='OUTPUT',
-                    nargs = 1,
-                    help='Path to the output file.')
+input_files = snakemake.input.edf
+annot_files = snakemake.input.annot
+tsv_file = snakemake.input.tsv
+processes = int(snakemake.params.processes)
+output_dir = snakemake.output.out_dir
+parc_path = snakemake.input.parc
+noncon_to_con_tf_path = snakemake.input.tf
 
-args = parser.parse_args()
-
-input_files = args.input
-annot_files = args.annot
-output_dir = args.output[0]
-processes = int(args.processes[0])
-tsv_file = args.tsv[0]
 print(tsv_file)
 print(output_dir)
 # Import file:
@@ -69,25 +52,45 @@ try:
             f.close()
             # Find time stamps where the 'awake trigger' event is happening
             time_stamps = onset_list[id]
-            # Copy file to local scratch
-            shutil.copy(edf, '/tmp/')
-            new_edf ='/tmp/'+re.search(pattern_complete, edf).group()
-            # Build output folder/file
+            # Size of file in MB
+            size_edf = os.path.getsize(edf)/1000000
+            # Copy file to local scratch if possible
+            new_edf = None
+            if size_edf <= 0.9*int(snakemake.params.mem):
+                # Copy file to local scratch
+                print('caca2')
+                shutil.copy(edf, '/tmp/')
+                new_edf ='/tmp/'+re.search(pattern_complete, edf).group()
+            # Build output folder/file for edf file and tsv
             output_path_ieeg = re.search(pattern_dir, edf).group()
             out_file_name = re.search(pattern=pattern_ieeg, string=edf).group()
+            out_tsv = re.search(r'sub-\d{3}', edf).group()+'_space-native_SEEGA.tsv'
             if output_dir.endswith('/'):
                 output_path_ieeg = output_dir + output_path_ieeg
                 os.makedirs(output_path_ieeg)
                 out_path_name = output_path_ieeg+out_file_name+'_epoch.edf'
+                out_tsv = output_dir + out_tsv
             else:
                 output_path_ieeg = output_dir+'/'+output_path_ieeg
                 os.makedirs(output_path_ieeg)
                 out_path_name = output_path_ieeg+out_file_name+'_epoch.edf'
+                out_tsv = output_dir + '/' + out_tsv
+
             # Extract info about electrodes positions
             elec_pos = pd.read_csv(tsv_file, sep='\t')
+            # Create bipolar combinations
+            bipolar_channels, bipolar_info_df = create_bipolars(elec_pos, processes)
+            # Create tsv file with information about bipolar channels
+            print('here')
+            create_bipolar_tsv(parc_path, noncon_to_con_tf_path, bipolar_info_df, out_tsv)
             # Here call function to create new EDF file
-            create_EDF(new_edf, time_stamps, elec_pos, out_path_name, processes)
-            os.remove(new_edf)
+            if (new_edf == None):
+                create_EDF(edf, time_stamps, bipolar_channels, out_path_name, processes)
+            else:
+                print('aqui')
+                create_EDF(new_edf, time_stamps, bipolar_channels, out_path_name, processes)
+                print('delete')
+                os.remove(new_edf)
 except FileNotFoundError:
     raise FileNotFoundError('The *.edf file was not found in the indicated \
         path.')
